% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11omp-package.R
\name{squared_named}
\alias{squared_named}
\title{Named list with squared numbers and the threads used}
\usage{
squared_named(x)
}
\arguments{
\item{x}{A vector of doubles}
}
\description{
Named list with squared numbers and the threads used
}
