% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11omp-package.R
\name{squared_unnamed}
\alias{squared_unnamed}
\title{Unnamed list with squared numbers and the threads used}
\usage{
squared_unnamed(x)
}
\arguments{
\item{x}{A vector of doubles}
}
\description{
Unnamed list with squared numbers and the threads used
}
